# Cache compiled get and fill list methods in Deserializer