This adds caching of methods retrieved from `GetAndFillListMethod` in a `ConcurrentDictionary`, making deserialization of lists a lot faster for multiple runs.

Benchmark with serialization of pinit packet with two members in a list
```
| Method |       Mean |     Error |     StdDev |
| Before | 410.157 us | 8.1389 us | 12.4290 us |
|  After | 120.962 us | 2.4134 us |  5.2465 us |
```