@0Lucifer0 2022-01-02
Thanks for the contribution LGTM
Will be in the next release